Removed usage of dynamic_cast
